requirements.txt
Flask
Flask-CORS
python-dotenv
youtube-transcript-api==0.6.2
google-generativeai==0.8.3
gunicorn